aiffwrite.c:   Calculate the SSND chunk size correctly even when using compressed data   formats. wave.c, wave.h, wavewrite.c:   Fix problems with WAVE files which have compressed data. testfloat.c:   Print the name of the file format being tested. writeulaw.c, writealaw.c:   Update writeulaw and writealaw to test AIFF-C and WAVE in addition to   NeXT .snd. ChangeLog:   Add entries for today's activities. util.h, units.h, util.c:   Make _af_format_sample_size and _af_format_frame_size consider compression.
BKrev: 3b4a67f5idNuusOYjTXuqPPlAg5p4A

+++ libaudiofile/util.h
@@ -57,9 +57,9 @@ _Track *_af_filehandle_get_track (AFfilehandle file, int trackid);
 
 bool _af_unique_ids (int *ids, int nids, char *idname, int iderr);
 
-int _af_format_frame_size (_AudioFormat *format, bool stretch3to4);
+float _af_format_frame_size (_AudioFormat *format, bool stretch3to4);
 int _af_format_frame_size_uncompressed (_AudioFormat *format, bool stretch3to4);
-int _af_format_sample_size (_AudioFormat *format, bool stretch3to4);
+float _af_format_sample_size (_AudioFormat *format, bool stretch3to4);
 int _af_format_sample_size_uncompressed (_AudioFormat *format, bool stretch3to4);
 
 status _af_set_sample_format (_AudioFormat *f, int sampleFormat, int sampleWidth);
